% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purl_chunks.R
\name{purl_chunks}
\alias{purl_chunks}
\title{purl_chunks}
\usage{
purl_chunks(input_file, input_path = "src/chunks")
}
\arguments{
\item{input_file}{String of the input file name to the Rmarkdown file that will be purlled into seperate chunks}

\item{input_path}{String of the path where to place the R files of each of the chunks}
}
\description{
The function purl_chunks takes an Rmarkdown file and returns each chunk as a
seperate R file.  It was found on stackoverflow at
(https://stackoverflow.com/questions/35855837/with-knitr-preserve-chunk-options-when-purling-chunks-into-separate-files)
}
\seealso{
knitr
}
\author{
Kirk Gosik <kdgosik@gmail.com>
}
\keyword{knitr}
\keyword{purl}
