% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_site.R
\name{update_site}
\alias{update_site}
\title{update_site}
\usage{
update_site()
}
\description{
The function update_site is meant to make reporducible workflows easier and more standardized.  This will
update the R files that are in the src/chunks directory as well as knit new html files for the Notebook file.
}
\seealso{
knitr
}
\author{
Kirk Gosik <kdgosik@gmail.com>
}
\keyword{knitr}
\keyword{purl}
